fix: extract the right-click menu as a component
